fix: surface and resolve warnings only visible under -Werror (#23 Phase 3 fixup)
Phase 1b's sample build cleared 108 warnings from the most-touched
TUs, but several categories only showed up once -Werror forced the
full LTO build to drag every TU through the analysis. This commit
catches up:

- Add the missing POW_SCALAR case in
  src/models/{behavioral,transactional}/compute/compute_fabric.cpp
  (was caught by gcc's -Wswitch and was the first CI failure).

- Add {} default-member-initializer to TraceEvent::loop_state in
  include/sw/kpu/behavioral/tiled_matmul_program.hpp - eliminates 17
  -Wmissing-field-initializers sites at the aggregate brace-init
  call sites.

- Mark Universal third-party include as SYSTEM at all four call sites
  (kpu_behavioral, mlp_${NAME}, quantization_tests, quantized_mlp_tests).
  Universal's decimal.hpp triggers -Wignored-qualifiers in some
  templates; SYSTEM tells gcc to leave third-party headers alone.

- ~30 [[maybe_unused]] additions across tests/noc, tests/compute,
  include/sw/kpu/harness, include/sw/analysis, src/software/isa,
  tools/calibration, tools/dfg, and patterns/memory/{gddr6,hbm2,
  hbm3,lpddr5}/. These are TUs the earlier Phase 1b sweep didn't
  rebuild because my touched-header sample didn't propagate to them.

- Cast int -> uint64_t in four 'expected_reads = ACCESSES_PER_CHANNEL'
  comparisons in patterns/memory/hbm2,hbm3/multi-channel/ to clear
  -Wsign-compare.

- Cast in tests/noc/wormhole_router_test.cpp:86 for the same reason.

Verified locally: full cmake -DCMAKE_CXX_FLAGS=-Werror build green;
test_memory_controller 12/12, c_stationary_matmul_test 14/14, and
vector_engine_test 12/12 all pass.

Refs #23

+++ patterns/memory/hbm2/multi-channel/hbm2_four_channel.cpp
@@ -54,7 +54,7 @@ int main() {
     }
 
     const auto& stats = harness.stats();
-    int expected_reads = ACCESSES_PER_CHANNEL * NUM_CHANNELS;
+    uint64_t expected_reads = ACCESSES_PER_CHANNEL * NUM_CHANNELS;
     if (stats.reads != expected_reads) {
         std::cerr << "FAIL: Expected " << expected_reads << " reads, got " << stats.reads << std::endl;
         return 1;
